Move addWalletFormStore interface
Move addWalletFormStore interface from wallets/index.tsx to src/interface/wallets since interface itself is used all around the app.
Fix imports due to moving the interface.
